pyproject.toml
[project]
name = "invariant-server"
version = "0.1.0"
description = "Invariant policy server"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "invariant @ git+https://github.com/invariantlabs-ai/invariant.git",
    "pydantic-settings>=2.3.4",
    ### Invariant extra dependencies
    "presidio-analyzer>=2.2.354",
    "spacy>=3.7.5",
    "fastapi-utils>=0.7.0",
    "typing-inspect>=0.9.0",
    "huggingface-hub>=0.24.5",
    "transformers>=4.44.0",
    "torch>=2.4.0",
    "numpy<2",
    "fastapi-cache2>=0.2.2",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.2.2",
    "requests>=2.32.3",
    "ruff>=0.5.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["server"]
